requirements.txt
qwen-tts
soundfile
pyyaml
numpy
requests
fugashi
unidic-lite
